+++ Changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+* Show a debug message if the settings schema could not be loaded
 * Fixed division by zero error
 
 ## v0.3.0 (2017-09-04)